requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
